Refactor recipe.py
